Edit: moved selectedGame repositories to domain layer and factories to infra layer
