+++ src/modules/selectedGame/domain/repositories/IActivitiesRepository.ts
