+++ src/modules/selectedGame/infra/repositories/AchievementsRepository.ts
@@ -1,5 +1,5 @@
 import IAchievement from 'modules/selectedGame/domain/entities/IAchievement';
-import IAchievementsRepository from 'modules/selectedGame/repositories/IAchievementsRepository';
+import IAchievementsRepository from 'modules/selectedGame/domain/repositories/IAchievementsRepository';
 import makeHttpProvider from 'shared/infra/providers/factories/makeHttpProvider';
 
 export default class AchievementsRepository implements IAchievementsRepository {
