+++ src/modules/selectedGame/infra/repositories/ActivitiesRepository.ts
@@ -1,5 +1,5 @@
 import IActivity from 'modules/selectedGame/domain/entities/IActivity';
-import IActivitiesRepository from 'modules/selectedGame/repositories/IActivitiesRepository';
+import IActivitiesRepository from 'modules/selectedGame/domain/repositories/IActivitiesRepository';
 import makeHttpProvider from 'shared/infra/providers/factories/makeHttpProvider';
 
 export default class ActivitiesRepository implements IActivitiesRepository {
