+++ src/modules/selectedGame/infra/repositories/FeedPostsRepository.ts
@@ -1,5 +1,5 @@
 import IFeedPost from 'modules/selectedGame/domain/entities/IFeedPost';
-import IFeedPostsRepository from 'modules/selectedGame/repositories/IFeedPostsRepository';
+import IFeedPostsRepository from 'modules/selectedGame/domain/repositories/IFeedPostsRepository';
 import makeHttpProvider from 'shared/infra/providers/factories/makeHttpProvider';
 
 export default class FeedPostsRepository implements IFeedPostsRepository {
