+++ src/modules/selectedGame/infra/repositories/factories/makeAchievementsRepository.ts
@@ -1,5 +1,5 @@
+import IAchievementsRepository from 'modules/selectedGame/domain/repositories/IAchievementsRepository';
 import AchievementsRepository from 'modules/selectedGame/infra/repositories/AchievementsRepository';
-import IAchievementsRepository from 'modules/selectedGame/repositories/IAchievementsRepository';
 
 export default function makeAchievementsRepository(): IAchievementsRepository {
   return new AchievementsRepository();
