+++ src/modules/selectedGame/infra/repositories/factories/makeActivitiesRepository.ts
@@ -1,5 +1,5 @@
+import IActivitiesRepository from 'modules/selectedGame/domain/repositories/IActivitiesRepository';
 import ActivitiesRepository from 'modules/selectedGame/infra/repositories/ActivitiesRepository';
-import IActivitiesRepository from 'modules/selectedGame/repositories/IActivitiesRepository';
 
 export default function makeActivitiesRepository(): IActivitiesRepository {
   return new ActivitiesRepository();
