+++ src/modules/selectedGame/infra/repositories/factories/makeFeedPostsRepository.ts
@@ -1,5 +1,5 @@
+import IFeedPostsRepository from 'modules/selectedGame/domain/repositories/IFeedPostsRepository';
 import FeedPostsRepository from 'modules/selectedGame/infra/repositories/FeedPostsRepository';
-import IFeedPostsRepository from 'modules/selectedGame/repositories/IFeedPostsRepository';
 
 export default function makeFeedPostsRepository(): IFeedPostsRepository {
   return new FeedPostsRepository();
