+++ src/modules/selectedGame/infra/repositories/factories/makeLeaderboardsRepository.ts
@@ -1,5 +1,5 @@
+import ILeaderboardsRepository from 'modules/selectedGame/domain/repositories/ILeaderboardsRepository';
 import LeaderboardsRepository from 'modules/selectedGame/infra/repositories/LeaderboardsRepository';
-import ILeaderboardsRepository from 'modules/selectedGame/repositories/ILeaderboardsRepository';
 
 export default function makeLeaderboardsRepository(): ILeaderboardsRepository {
   return new LeaderboardsRepository();
