+++ src/modules/selectedGame/infra/repositories/factories/makePlayersRepository.ts
@@ -1,5 +1,5 @@
+import IPlayersRepository from 'modules/selectedGame/domain/repositories/IPlayersRepository';
 import PlayersRepository from 'modules/selectedGame/infra/repositories/PlayersRepository';
-import IPlayersRepository from 'modules/selectedGame/repositories/IPlayersRepository';
 
 export default function makePlayersRepository(): IPlayersRepository {
   return new PlayersRepository();
