+++ src/modules/selectedGame/infra/repositories/factories/makeRequestsRepository.ts
@@ -1,5 +1,5 @@
+import IRequestsRepository from 'modules/selectedGame/domain/repositories/IRequestsRepository';
 import RequestsRepository from 'modules/selectedGame/infra/repositories/RequestsRepository';
-import IRequestsRepository from 'modules/selectedGame/repositories/IRequestsRepository';
 
 export default function makeRequestsRepository(): IRequestsRepository {
   return new RequestsRepository();
