+++ src/modules/selectedGame/services/GetAchievementsService.ts
@@ -1,5 +1,5 @@
 import IAchievement from 'modules/selectedGame/domain/entities/IAchievement';
-import IAchievementsRepository from 'modules/selectedGame/repositories/IAchievementsRepository';
+import IAchievementsRepository from 'modules/selectedGame/domain/repositories/IAchievementsRepository';
 
 interface IExecute {
   achievements?: IAchievement[];
