+++ src/modules/selectedGame/services/factories/makeFindPlayerService.ts
@@ -1,4 +1,4 @@
-import makePlayersRepository from 'modules/selectedGame/repositories/factories/makePlayersRepository';
+import makePlayersRepository from 'modules/selectedGame/infra/repositories/factories/makePlayersRepository';
 import FindPlayerService from 'modules/selectedGame/services/FindPlayerService';
 
 export default function makeFindPlayerService(): FindPlayerService {
