+++ src/modules/selectedGame/services/factories/makeGetAchievementsService.ts
@@ -1,4 +1,4 @@
-import makeAchievementsRepository from 'modules/selectedGame/repositories/factories/makeAchievementsRepository';
+import makeAchievementsRepository from 'modules/selectedGame/infra/repositories/factories/makeAchievementsRepository';
 import GetAchievementsService from 'modules/selectedGame/services/GetAchievementsService';
 
 export default function makeGetAchievementsService(): GetAchievementsService {
