+++ src/modules/selectedGame/services/factories/makeGetCurrentLeaderboard.ts
@@ -1,4 +1,4 @@
-import makeLeaderboardsRepository from 'modules/selectedGame/repositories/factories/makeLeaderboardsRepository';
+import makeLeaderboardsRepository from 'modules/selectedGame/infra/repositories/factories/makeLeaderboardsRepository';
 import GetCurrentLeaderboard from 'modules/selectedGame/services/GetCurrentLeaderboard';
 
 export default function makeGetCurrentLeaderboard(): GetCurrentLeaderboard {
