+++ src/modules/selectedGame/services/factories/makeGetFeedPostsService.ts
@@ -1,4 +1,4 @@
-import makeFeedPostsRepository from 'modules/selectedGame/repositories/factories/makeFeedPostsRepository';
+import makeFeedPostsRepository from 'modules/selectedGame/infra/repositories/factories/makeFeedPostsRepository';
 import GetFeedPostsService from 'modules/selectedGame/services/GetFeedPostsService';
 
 export default function makeGetFeedPostsService(): GetFeedPostsService {
