+++ src/modules/selectedGame/services/factories/makeRequestAchievementUnlockService.ts
@@ -1,4 +1,4 @@
-import makeRequestsRepository from 'modules/selectedGame/repositories/factories/makeRequestsRepository';
+import makeRequestsRepository from 'modules/selectedGame/infra/repositories/factories/makeRequestsRepository';
 import RequestAchievementUnlockService from 'modules/selectedGame/services/RequestAchievementUnlockService';
 
 export default function makeRequestAchievementUnlockService(): RequestAchievementUnlockService {
